TST: Update complex long double precision tests.
It seems that special casing the long double precision tests in order to
deal with PPC is no longer needed now that finfo has been fixed for the
IBM double double and the blacklisting of failing functions.
Consequently the tests can be simplified and no longer needed functions
removed. See #8566 for discussion.