BUG: Fix for #10533 np.dtype(ctype) does not respect endianness (#12380)
* BUG: Fix for #10533 np.dtype(ctype) does not respect endianness

See: #10533
Added some code that uses the __name__ of the ctypes type. Also
added some new test to make sure we'll know if the ctypes data
that allows this workaround breaks in the future.

* Fix for #10533 np.dtype(ctype) does not respect endianness

Removed the TODO for this bug.

* Fix for #10533 np.dtype(ctype) does not respect endianness

Solved the recursiveness issue that failed the automated test.
Also did some of Eric's requested changes.

* Small mistake in the new function fixed.

* Changed 'if t.__name__.endswith('_be') ' with
'if t.__ctypes_be__ is t' on Eric's suggestion.

* Code cleanup on Eric's recommendations.

* Cleaned up the tests on Eric's suggestions.